+++ jest.config.js
@@ -12,6 +12,7 @@ module.exports = {
     "fixtures",
     "componentHelpers",
   ],
+  setupFiles: ["./jest.setEnv.ts"],
   setupFilesAfterEnv: ["./jest.setup.ts"],
   snapshotSerializers: ["enzyme-to-json/serializer"],
   resetMocks: true,