+++ jest.setEnv.ts
@@ -1,0 +1,3 @@
+process.env.NEXT_PUBLIC_READER_VERSION = "v2";
+
+export {};