+++ next.config.js
@@ -1,6 +1,7 @@
 const withTM = require("next-transpile-modules")([
   "library-simplified-webpub-viewer",
 ]);
+
 const path = require("path");
 
 const config = {
@@ -9,6 +10,7 @@ const config = {
   },
   webpack: (config) => {
     config.resolve.alias["~"] = path.resolve(__dirname);
+
     return config;
   },
   sassOptions: {