+++ src/constants/mediaTypes.ts
@@ -1,4 +1,3 @@
-//Read should have `application/pdf` as well but the reader doesn't support it yet
 export const MediaTypes = {
   read: ["application/epub+xml", "application/webpub+json"],
   download: ["application/pdf", "application/epub+zip"],