+++ src/util/__tests__/apiConversion.test.ts
@@ -1,5 +1,4 @@
 /* eslint-env mocha */
-import React from "react";
 import { SearchField } from "~/src/types/DataModel";
 import { ApiSearchQuery, SearchQuery } from "~/src/types/SearchQuery";
 import { toApiQuery, toSearchQuery } from "../apiConversion";
@@ -105,6 +104,7 @@ describe("Converting search query to api query", () => {
       showAll: "true",
       query: 'keyword:"Civil War" OR Lincoln,author:last, first',
       sort: "title:asc",
+      readerVersion: "v2",
     };
 
     expect(toApiQuery(searchQuery)).toEqual(expectedApiQuery);
@@ -120,6 +120,7 @@ describe("Converting search query to api query", () => {
   test("converts searchQuery to apiQuery with minimal information", () => {
     const minimalApiQuery: ApiSearchQuery = {
       query: "keyword:cat",
+      readerVersion: "v2",
     };
     expect(
       toApiQuery({ queries: [{ field: SearchField.Keyword, query: "cat" }] })